add WorkspaceSupport, factor some code out of WSService
